Issue #2363 Add more vector configuration optons. Change-Id: I1a3e564629b9e4afe93a0a4ceee8e57681aa0644
Former-commit-id: ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2 [formerly ff06ea1c5af7caa4cc796c5b5cb3dd968474d4c2 [formerly 8af298f9ef2c1c03e6be690c07829d0569b1dedd]]
Former-commit-id: 20853f733378677864a5bbb08cd27f47a6646f12
Former-commit-id: 6c7c693581a84d7bce05b1e6f2b2901e4aa5985d

+++ cave/com.raytheon.viz.grid/src/com/raytheon/viz/grid/rsc/GriddedIconDisplay.java
@@ -47,11 +47,12 @@ import com.vividsolutions.jts.geom.Coordinate;
  * <pre>
  * 
  * SOFTWARE HISTORY
- * Date         Ticket#    Engineer    Description
- * ------------ ---------- ----------- --------------------------
- * Apr 23, 2010            bsteffen     Initial creation
- * Aug 27, 2013     #2287  randerso     Added densityFactor to allow application specific 
- *                                      adjustment of density.
+ * Date          Ticket#  Engineer    Description
+ * ------------- -------- ----------- --------------------------
+ * Apr 23, 2010           bsteffen    Initial creation
+ * Aug 27, 2013  2287     randerso    Added densityFactor to allow application
+ *                                    specific adjustment of density.
+ * Sep 23, 2013  2363     bsteffen    Add more vector configuration options.
  * 
  * </pre>
  * 
@@ -109,7 +110,7 @@ public class GriddedIconDisplay extends AbstractGriddedDisplay<IImage> {
     @Override
     protected IImage createResource(Coordinate coord) throws VizException {
         if (iconFactory == null) {
-            iconFactory = new PointIconFactory(color, size);
+            iconFactory = new PointIconFactory(color, (int) size);
         }
         int i = getValue(coord);
         IImage image = images.get(i);
